fix: Moved all types to the types.ts file
I was refactoring the code to allow for experiments without replacement and hit a type import cycle.  Instead of adding to the complexity of that PR this PR just centralizes all the types into one file to remove the import cycle.
