+++ src/models/device-model.tsx
@@ -1,57 +1,9 @@
-import { Id, createId } from "../utils/id";
-
-export enum ViewType {
-  Mixer = "mixer",
-  Spinner = "spinner",
-  Collector = "collector"
-}
-
-export type View = ViewType.Mixer | ViewType.Spinner | ViewType.Collector;
-
-export interface IDevice {
-  id: Id;
-  viewType: View;
-  variables: IVariables;
-  collectorVariables: ICollectorVariables;
-  formulas: Record<string, string>;
-}
-
-// a map of variables to their percentages
-// i.e.: { a: 50, b: 50 }
-// the view (mixer / spinner) decides how to best represent these percentages
-export type IVariables = Array<string>;
-// a map of attributes to their values
-// { "Mammal": "Dog", "Color": "Brown"}
-export interface ICollectorItem {
-  [attr: string]: string | number;
-}
-
-// an array of items that come from a linked dataset
-// i.e.: [ { "Mammal": "Dog", "Color": "Brown"}, { "Mammal": "Cat", "Color": "Black"} ]
-// the collector view will represent the values of the first key-value pair of each object in the array
-// i.e., one ball for "Dog", one ball for "Cat"
-// if "Dog" is selected by the collector, the entire item is sent to CODAP
-export type ICollectorVariables = Array<ICollectorItem>;
-export interface IDataContext {
-  guid: number;
-  id: number;
-  name: string;
-  title: string;
-}
-
-export interface IItem {
-  id: number;
-  values: ICollectorItem;
-}
-
-export type IItems = Array<IItem>;
-
-export type ClippingDef = { id: string, element: JSX.Element };
+import { IVariables, IDevice, ViewType } from "../types";
+import { createId } from "../utils/id";
 
 export function gcd(a: number, b: number): number {
   return b === 0 ? a : gcd(b, a % b);
 }
-
 
 export function lcm(a: number, b: number): number {
   return (a * b) / gcd(a, b);
